requirements.txt
requests
protobuf
pycryptodome
aiohttp
cfonts
python-telegram-bot
flask
gunicorn
google-api-python-client